쿼터 제한 초과 방지

+++ stelline/apis/search/search_api.py
@@ -209,6 +209,9 @@ def search_api_process(by_admin=False):
             searched_time = new_songs["searched_time"]
             conn = get_rds_connection()
             try:
+                if not all_songs:
+                    logging.info("새로운 검색 데이터 없음")
+                    continue
                 with conn.cursor() as cursor:
                     sql = """
                         TRUNCATE TABLE songs_data;
